+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/ui/MaterialPropertiesTable.kt
@@ -25,6 +25,8 @@ import com.intellij.ui.table.JBTable
 import it.czerwinski.intellij.wavefront.editor.model.MaterialProperty
 import it.czerwinski.intellij.wavefront.editor.model.materialProperties
 import it.czerwinski.intellij.wavefront.lang.psi.MtlIlluminationValueElement
+import it.czerwinski.intellij.wavefront.lang.psi.MtlReflectionType
+import it.czerwinski.intellij.wavefront.lang.psi.MtlScalarChannel
 import javax.swing.DefaultCellEditor
 import javax.swing.table.TableCellEditor
 import javax.swing.table.TableCellRenderer
@@ -52,6 +54,18 @@ class MaterialPropertiesTable(
 
     private val myTextureEditor = TextureTableCellEditor(project, parent = editor)
 
+    private val myScalarChannelEditor = DefaultCellEditor(
+        ComboBox((listOf(null) + enumValues<MtlScalarChannel>()).toTypedArray()).apply {
+            renderer = SimpleListCellRenderer.create("") { value -> value?.description }
+        }
+    )
+
+    private val myReflectionTypeEditor = DefaultCellEditor(
+        ComboBox((listOf(null) + enumValues<MtlReflectionType>()).toTypedArray()).apply {
+            renderer = SimpleListCellRenderer.create("") { value -> value?.description }
+        }
+    )
+
     override fun getCellRenderer(row: Int, column: Int): TableCellRenderer = myRenderer
 
     override fun getCellEditor(row: Int, column: Int): TableCellEditor = when (materialProperties[row]) {
@@ -59,6 +73,9 @@ class MaterialPropertiesTable(
         is MaterialProperty.MaterialIlluminationValue -> myIlluminationEditor
         is MaterialProperty.MaterialFloatValue -> myFloatEditor
         is MaterialProperty.MaterialTexture -> myTextureEditor
+        is MaterialProperty.MaterialTextureScalarChannel -> myScalarChannelEditor
         is MaterialProperty.MaterialTextureValueModifier -> myFloatEditor
+        is MaterialProperty.MaterialReflectionTexture -> myTextureEditor
+        is MaterialProperty.MaterialReflectionTextureType -> myReflectionTypeEditor
     }
 }