+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/ui/MaterialPropertiesTableCellRenderer.kt
@@ -20,6 +20,8 @@ import com.intellij.ui.components.JBLabel
 import com.intellij.util.ui.JBEmptyBorder
 import it.czerwinski.intellij.common.ui.ColorLabel
 import it.czerwinski.intellij.wavefront.lang.psi.MtlIlluminationValueElement
+import it.czerwinski.intellij.wavefront.lang.psi.MtlReflectionType
+import it.czerwinski.intellij.wavefront.lang.psi.MtlScalarChannel
 import java.awt.Color
 import java.awt.Component
 import javax.swing.JTable
@@ -40,6 +42,8 @@ class MaterialPropertiesTableCellRenderer : TableCellRenderer {
         is Float -> JBLabel(value.toString(), SwingConstants.TRAILING)
         is Color -> ColorLabel(value)
         is MtlIlluminationValueElement.Illumination -> JBLabel(value.description)
+        is MtlScalarChannel -> JBLabel(value.description)
+        is MtlReflectionType -> JBLabel(value.description)
         else -> JBLabel(value?.toString().orEmpty())
     }.apply {
         border = labelBorder