+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - Add support for platform version `231.*`
 - Add support for Apple M1 processor (macOS AArch64)
 - Automatically preview material at caret
-- Add support for scalar map channel option (`-imfchan`) in 3D preview (only RGB)
+- Add support for scalar map channel option (`-imfchan`) in 3D preview and material editor (only RGB)
 - Add support for spherical reflection map (`refl -type sphere`) in 3D preview
+- Add support for reflection map type option in material editor
 - Settings for MTL file editor layout separate from OBJ file editor layout
 - Setting for generating mipmaps for textures (on/off)
 - Setting for selecting face culling
@@ -51,6 +52,7 @@
 ### Fixed
 - Fix error in 3D preview of a line without texture coordinates
 - Wrap file operations in read-actions
+- Fix error when adding reflection texture via material editor
 - Replace usages of internal method `StartupManager.runAfterOpened()` with a `StartupActivity`
 - Override `getActionUpdateThread()` in existing implementations of `AnAction`
 - Use DPI-aware `JBEmptyBorder` instead of `EmptyBorder`